Refactor action types under src/screens/Users domain
- Move action types from action.js to types.js
- Update import statement's paths according to the changes
