+++ src/screens/Users/screens/Profile/data/profile/actions.js
@@ -1,7 +1,5 @@
 // Action types
-export const PROFILE_GET = 'Profile/data/PROFILE_GET';
-export const PROFILE_GET_FAILURE = 'Profile/data/PROFILE_GET_FAILURE';
-export const PROFILE_GET_SUCCESS = 'Profile/data/PROFILE_GET_SUCCESS';
+import { PROFILE_GET, PROFILE_GET_FAILURE, PROFILE_GET_SUCCESS } from './types';
 
 // Get user profile : Start
 export const getProfile = () => ({
