+++ src/screens/Users/screens/Profile/data/profile/reducers.js
@@ -3,7 +3,7 @@ import { createSelector } from 'reselect';
 
 // Action types
 import { USER_RESET } from 'data/session/types';
-import { PROFILE_GET, PROFILE_GET_FAILURE, PROFILE_GET_SUCCESS } from './actions';
+import { PROFILE_GET, PROFILE_GET_FAILURE, PROFILE_GET_SUCCESS } from './types';
 
 // Initial state
 const initialState = null;
