+++ src/screens/Users/screens/Profile/data/profile/sagas.js
@@ -8,8 +8,9 @@ import { getError } from 'helpers/state';
 // Services
 import * as usersService from 'services/users';
 
-// Action types and action creators
+// Action creators and action types
 import * as actions from './actions';
+import * as types from './types';
 
 // Get user profile
 function* getProfile(action) {
@@ -34,7 +35,7 @@ function* getProfile(action) {
 
 // Actions watcher
 function* watcher() {
-  yield takeLatest(actions.PROFILE_GET, getProfile);
+  yield takeLatest(types.PROFILE_GET, getProfile);
 }
 
 // Module exports
