+++ src/screens/Users/screens/Profile/data/profile/types.js
@@ -1,0 +1,3 @@
+export const PROFILE_GET = 'Profile/data/PROFILE_GET';
+export const PROFILE_GET_FAILURE = 'Profile/data/PROFILE_GET_FAILURE';
+export const PROFILE_GET_SUCCESS = 'Profile/data/PROFILE_GET_SUCCESS';
