+++ src/screens/Users/screens/Profile/reducers.js
@@ -25,7 +25,7 @@ import {
   USER_DELETE_SUCCESS,
   USER_RESET
 } from 'data/session/types';
-import { PROFILE_GET, PROFILE_GET_FAILURE, PROFILE_GET_SUCCESS } from './data/profile/actions';
+import { PROFILE_GET, PROFILE_GET_FAILURE, PROFILE_GET_SUCCESS } from './data/profile/types';
 
 // Reducers
 import data from './data/reducers';
