+++ src/screens/Users/screens/SignIn/actions.js
@@ -1,5 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+
 // Action types
-export const SIGNIN_RESET_UI = 'SignIn/SIGNIN_RESET_UI';
+import { SIGNIN_RESET_UI } from './types';
 
 // Reset UI state
 export const resetUI = () => ({
