+++ src/screens/Users/screens/SignIn/types.js
@@ -1,0 +1,3 @@
+/* eslint-disable import/prefer-default-export */
+
+export const SIGNIN_RESET_UI = 'SignIn/SIGNIN_RESET_UI';
