+++ src/screens/Users/screens/SignUp/types.js
@@ -1,0 +1,3 @@
+/* eslint-disable import/prefer-default-export */
+
+export const SIGNUP_RESET_UI = 'SignUp/SIGNUP_RESET_UI';
